🐛 Refactor openDualPlan function to set copies to amount
This commit refactors the openDualPlan function in openDual.js to set the copies variable to the provided amount. Previously, the copies variable was hardcoded to 1, but now it is dynamically set based on the amount parameter. This change improves the flexibility of the function and allows for creating dual investment orders with different numbers of copies.