✨ feat: Update BuyLowInfo component to use Redux store for openDuals state
This commit modifies the BuyLowInfo component in BuyLowInfo.jsx to utilize the openDuals state from the Redux store. It imports the `useSelector` hook from `react-redux` and retrieves the `openedBuyDuals` value from the store. This change ensures that the `openDuals` variable in the component is always up-to-date with the latest state, improving the accuracy of the displayed information.