🐛 refactor: Schedule listDualInvestmentPlans function to run recursively every 5 seconds
This commit updates the cron job in the Backend/index.js file to schedule the listDualInvestmentPlans function to run recursively every 5 seconds. By using a recursive approach, the function will be executed repeatedly at the specified interval, ensuring that the list of dual investment plans is updated regularly. This change improves the functionality and reliability of the application.