✨ feat: Add route for adding a short bot
Add a new route "/add-short-bot" to handle the addition of a short bot. This route expects the user ID, currency, strike price, and investment amount as parameters. If any of these fields are missing, a 400 error is returned. Upon successful addition of the short bot, a success message along with the newly created short bot data is returned. In case of any errors, a 500 error is returned.